core/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0
django-celery-results==2.5.1
django-filter==23.3
django-formtools==2.5.1
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-two-factor-auth==1.15.5
idna==3.6
kavenegar==1.1.2
kombu==5.3.6
phonenumbers==8.13.31
Pillow==10.1.0
prompt-toolkit==3.0.43
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
qrcode==7.4.2
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
black 
